+++ packages/res/CHANGELOG.md
@@ -1,6 +1,15 @@
 # @tinyhttp/res
 
+## 1.0.8
+
+### Patch Changes
+
+- Fix minor sendFile and download bugs
+- Updated dependencies [undefined]
+  - @tinyhttp/send@1.0.8
+
 ## 1.0.7
+
 ### Patch Changes
 
 - regenerate changelogs