+++ packages/send/CHANGELOG.md
@@ -1,6 +1,13 @@
 # @tinyhttp/send
 
+## 1.0.8
+
+### Patch Changes
+
+- Fix minor sendFile and download bugs
+
 ## 1.0.7
+
 ### Patch Changes
 
 - regenerate changelogs